[fix] reduce function calls and clean up asap

+++ pagelet.js
@@ -91,11 +91,9 @@ Pagelet.prototype.configure = function configure(name, data, roots) {
   // However do not destroy assets as unauthorized pagelets won't register
   // assets in the first place and they might be used by other pagelets.
   //
-  if (data.remove) return pagelet.once('finished', function destroy() {
-    return pagelet.destroy({
-      assets: false,
-      remove: true
-    });
+  if (data.remove) return pagelet.destroy({
+    assets: false,
+    remove: true
   });
 
   //
